breakout bot functions into inidividual files
